+++ CHANGELOG.md
@@ -258,6 +258,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - **CUMULUS-2507**
     - Updated granule record write logic to set granule status to `failed` in both Postgres and DynamoDB if any/all of its files fail to write to the database.
 
+- **CUMULUS-2474**
+  - Add `S3ObjectStore` to `aws-client`
+
 ### Deprecated
 
 - **CUMULUS-2185** - RDS Migration Epic
@@ -276,9 +279,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Published `@cumulus/api@7.2.1-alpha.0` for dashboard testing
 - **CUMULUS-2469**
   - Added `tf-modules/cumulus_distribution` module to standup a skeleton distribution api
-
-- **CUMULUS-2474**
-  - Add `S3ObjectStore` to `aws-client`
 
 ## [v8.0.0] 2021-04-08