+++ CHANGELOG.md
@@ -43,7 +43,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
       accessible via the Core API.
   
 - **CUMULUS-2488**
-  - Removed all EMS lambdas, endpoints, params, etc
+  - Removed all EMS reporting including lambdas, endpoints, params, etc as all
+    reporting is now handled through Cloud Metrics
 
 ### Added